+++ crates/shipper-core/src/engine/mod.rs
@@ -32,9 +32,16 @@ use crate::webhook::{self, WebhookEvent};
 
 mod preflight;
 mod publish;
+mod readiness;
 mod rehearsal;
+mod retry;
 
 pub use preflight::PreflightRunOptions;
+use readiness::verify_published;
+use retry::{
+    emit_retry_backoff_event, record_rate_limit_observed_event, record_retry_backoff_event,
+    wait_after_retry,
+};
 
 pub trait Reporter {
     fn info(&mut self, msg: &str);
@@ -1466,256 +1473,6 @@ fn sequential_reconcile(
             Vec::new(),
         ),
     }
-}
-
-/// Emit a [`EventType::RetryBackoffStarted`] event + a human-readable warn
-/// line, then `thread::sleep(delay)`. Used at every retry-backoff site in the
-/// sequential publish loop so operators never stare at a silent CI log during
-/// the wait window. See #91. (The parallel path has a mirror helper in
-/// `engine::parallel::publish::emit_retry_backoff` that handles its
-/// `Arc<Mutex<_>>` wrapping.)
-#[allow(clippy::too_many_arguments)]
-fn emit_retry_backoff_event(
-    event_log: &mut events::EventLog,
-    events_path: &Path,
-    reporter: &mut dyn Reporter,
-    pkg_label: &str,
-    pkg_name: &str,
-    pkg_version: &str,
-    attempt: u32,
-    max_attempts: u32,
-    delay: std::time::Duration,
-    next_attempt_at: chrono::DateTime<Utc>,
-    reason: ErrorClass,
-    message: &str,
-) -> Result<()> {
-    record_retry_backoff_event(
-        event_log,
-        events_path,
-        pkg_label,
-        attempt,
-        max_attempts,
-        delay,
-        next_attempt_at,
-        &reason,
-        message,
-    )?;
-    wait_after_retry(
-        reporter,
-        pkg_name,
-        pkg_version,
-        attempt,
-        max_attempts,
-        delay,
-        reason,
-        message,
-    );
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-fn record_retry_backoff_event(
-    event_log: &mut events::EventLog,
-    events_path: &Path,
-    pkg_label: &str,
-    attempt: u32,
-    max_attempts: u32,
-    delay: std::time::Duration,
-    next_attempt_at: chrono::DateTime<Utc>,
-    reason: &ErrorClass,
-    message: &str,
-) -> Result<()> {
-    event_log.record(PublishEvent {
-        timestamp: Utc::now(),
-        event_type: EventType::RetryScheduled {
-            attempt,
-            max_attempts,
-            delay_ms: delay.as_millis() as u64,
-            next_attempt_at,
-            reason: reason.clone(),
-            message: message.to_string(),
-        },
-        package: pkg_label.to_string(),
-    });
-    record_publish_wait_event(
-        event_log,
-        events_path,
-        pkg_label,
-        delay,
-        "retry backoff",
-        Some(next_attempt_at),
-    )?;
-    event_log.record(PublishEvent {
-        timestamp: Utc::now(),
-        event_type: EventType::RetryBackoffStarted {
-            attempt,
-            max_attempts,
-            delay_ms: delay.as_millis() as u64,
-            next_attempt_at,
-            reason: reason.clone(),
-            message: message.to_string(),
-        },
-        package: pkg_label.to_string(),
-    });
-    event_log.write_to_file(events_path)?;
-    event_log.clear();
-    Ok(())
-}
-
-fn record_rate_limit_observed_event(
-    event_log: &mut events::EventLog,
-    events_path: &Path,
-    pkg_label: &str,
-    is_new_crate: bool,
-    retry_after: Option<std::time::Duration>,
-    message: &str,
-) -> Result<()> {
-    event_log.record(PublishEvent {
-        timestamp: Utc::now(),
-        event_type: EventType::RateLimitObserved {
-            is_new_crate,
-            retry_after_ms: retry_after.map(|delay| delay.as_millis() as u64),
-            message: message.to_string(),
-        },
-        package: pkg_label.to_string(),
-    });
-    event_log.write_to_file(events_path)?;
-    event_log.clear();
-    Ok(())
-}
-
-fn record_publish_wait_event(
-    event_log: &mut events::EventLog,
-    events_path: &Path,
-    pkg_label: &str,
-    delay: std::time::Duration,
-    reason: &str,
-    until: Option<chrono::DateTime<Utc>>,
-) -> Result<()> {
-    event_log.record(PublishEvent {
-        timestamp: Utc::now(),
-        event_type: EventType::PublishWaiting {
-            reason: reason.to_string(),
-            delay_ms: delay.as_millis() as u64,
-            until: until.unwrap_or_else(|| retry_next_attempt_at(delay)),
-        },
-        package: pkg_label.to_string(),
-    });
-    event_log.write_to_file(events_path)?;
-    event_log.clear();
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-fn wait_after_retry(
-    reporter: &mut dyn Reporter,
-    pkg_name: &str,
-    pkg_version: &str,
-    attempt: u32,
-    max_attempts: u32,
-    delay: std::time::Duration,
-    reason: ErrorClass,
-    message: &str,
-) {
-    // Delegate the human-facing narration AND the backoff sleep to the
-    // Reporter so TTY-capable UIs can render a live countdown during the
-    // wait. The default `Reporter::retry_wait` impl preserves the pre-#103
-    // behavior (single warn line + `thread::sleep(delay)`), so reporters
-    // that don't override it behave exactly as before.
-    reporter.retry_wait(
-        pkg_name,
-        pkg_version,
-        attempt,
-        max_attempts,
-        delay,
-        reason,
-        message,
-    );
-}
-
-fn verify_published(
-    reg: &RegistryClient,
-    crate_name: &str,
-    version: &str,
-    config: &crate::types::ReadinessConfig,
-    reporter: &mut dyn Reporter,
-    event_log: &mut events::EventLog,
-    events_path: &Path,
-    pkg_label: &str,
-) -> Result<(bool, Vec<ReadinessEvidence>)> {
-    reporter.info(&format!(
-        "{}@{}: readiness check ({:?})...",
-        crate_name, version, config.method
-    ));
-    let started_at = Instant::now();
-    record_readiness_event(
-        event_log,
-        events_path,
-        PublishEvent {
-            timestamp: Utc::now(),
-            event_type: EventType::ReadinessStarted {
-                method: config.method,
-            },
-            package: pkg_label.to_string(),
-        },
-    )?;
-    let mut emit_event = |event| record_readiness_event(event_log, events_path, event);
-    let (visible, evidence) = reg.is_version_visible_with_backoff_and_events(
-        crate_name,
-        version,
-        config,
-        &mut emit_event,
-    )?;
-    if visible {
-        reporter.info(&format!(
-            "{}@{}: visible after {} checks",
-            crate_name,
-            version,
-            evidence.len()
-        ));
-        record_readiness_event(
-            event_log,
-            events_path,
-            PublishEvent {
-                timestamp: Utc::now(),
-                event_type: EventType::ReadinessComplete {
-                    duration_ms: started_at.elapsed().as_millis() as u64,
-                    attempts: evidence.len() as u32,
-                },
-                package: pkg_label.to_string(),
-            },
-        )?;
-    } else {
-        reporter.warn(&format!(
-            "{}@{}: not visible after {} checks",
-            crate_name,
-            version,
-            evidence.len()
-        ));
-        record_readiness_event(
-            event_log,
-            events_path,
-            PublishEvent {
-                timestamp: Utc::now(),
-                event_type: EventType::ReadinessTimeout {
-                    max_wait_ms: config.max_total_wait.as_millis() as u64,
-                },
-                package: pkg_label.to_string(),
-            },
-        )?;
-    }
-    Ok((visible, evidence))
-}
-
-fn record_readiness_event(
-    event_log: &mut events::EventLog,
-    events_path: &Path,
-    event: PublishEvent,
-) -> Result<()> {
-    event_log.record(event);
-    event_log.write_to_file(events_path)?;
-    event_log.clear();
-    Ok(())
 }
 
 #[cfg(test)]
